requirements.txt
# Core Dependencies
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# HTTP Clients and API Integration
httpx==0.25.2
tenacity==8.2.3

# CLI Framework
click==8.1.7

# Git Integration (Optional)
GitPython==3.1.40
PyGithub==1.59.1

# Web Service Dependencies (Optional)
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Documentation (Optional)
mkdocs==1.5.3
mkdocs-material==9.4.8

# Monitoring and Logging
structlog==23.2.0